requirements.txt
pathlib
selenium
slack_sdk
python-dotenv
